Linux build fix

+++ C/src/ast.c
@@ -818,7 +818,7 @@ ast_print_type(Light_Type* type, u32 flags, s32 indent_level) {
             length += fprintf(out, "%.*s", type->alias.name->length, type->alias.name->data);
         } break;
         case TYPE_KIND_ARRAY:{
-            length += fprintf(out, "[%llu]", type->array_info.dimension);
+            length += fprintf(out, "[%llu]", (unsigned long long int)type->array_info.dimension);
             length += ast_print_type(type->array_info.array_of, flags, indent_level);
         } break;
         case TYPE_KIND_FUNCTION:{
